Inserting Google Analytics tracking to .html files that are not simple redirects.
(See review in https://github.com/google/blockly/pull/1282)

This commit branches off the last master release. The branch can be
rebased off of future master push commits as a means to insert the
analytics prior to updating the blockly-demo appspot domain.

When doing so, verify there are no new .html files with the following
command:

$ find . -type f -name "*.html" | xargs grep -L GoogleAnalyticsObject
./appengine/redirect.html
./demos/interpreter/index.html